refactor: split deploy effect version checks
