+++ src/core/deploy/execution.rs
@@ -10,6 +10,7 @@ use crate::core::extension::build::resolve_artifact_path_from_root;
 use crate::core::git;
 use crate::core::project::Project;
 
+use super::effect::remote_version_after_deploy_effect;
 use super::generated_artifacts::GeneratedBuildArtifactCleanupGuard;
 use super::path_roots::{component_remote_path, resolve_effective_remote_path};
 use super::planning::{calculate_directory_size, format_bytes};
@@ -18,8 +19,8 @@ use super::release_download;
 use super::safety_and_artifact::{deploy_artifact, deploy_via_git};
 use super::types::{ComponentDeployResult, DeployConfig, DeployResult};
 use super::version_overrides::{
-    deploy_with_override, fetch_remote_versions_for_project, find_deploy_override,
-    find_deploy_verification, is_self_deploy, prefer_installed_binary, run_post_deploy_hooks,
+    deploy_with_override, find_deploy_override, find_deploy_verification, is_self_deploy,
+    prefer_installed_binary, run_post_deploy_hooks,
 };
 
 pub(super) struct PreparedComponentDeploy {
@@ -883,51 +884,6 @@ fn execute_artifact_deploy(
     }
 }
 
-fn remote_version_after_deploy_effect(
-    component: &Component,
-    project: &Project,
-    base_path: &str,
-    client: &crate::core::server::SshClient,
-    effect: Option<&super::types::DeployEffect>,
-    local_version: Option<&String>,
-) -> std::result::Result<Option<String>, String> {
-    let Some(effect) = effect else {
-        return Ok(local_version.cloned());
-    };
-    let has_version_targets = component
-        .version_targets
-        .as_ref()
-        .is_some_and(|targets| !targets.is_empty());
-
-    if !has_version_targets {
-        return Ok(local_version.cloned());
-    }
-
-    let observed_versions = fetch_remote_versions_for_project(
-        std::slice::from_ref(component),
-        Some(project),
-        base_path,
-        client,
-    );
-    let observed = observed_versions.get(&component.id).cloned().ok_or_else(|| {
-        format!(
-            "Deploy command completed for '{}' at '{}', but Homeboy could not read remote_version from the applied tree. Refusing to report success from stale pre-deploy observations.",
-            component.id, effect.remote_path
-        )
-    })?;
-
-    if let Some(expected) = local_version {
-        if &observed != expected {
-            return Err(format!(
-                "Deploy command completed for '{}' at '{}', but post-deploy remote_version is '{}' instead of expected local_version '{}'. Refusing to report success from a mismatched deploy effect.",
-                component.id, effect.remote_path, observed, expected
-            ));
-        }
-    }
-
-    Ok(Some(observed))
-}
-
 /// Remove the local build artifact created for a deploy.
 ///
 /// `homeboy build` intentionally leaves a package for humans to inspect. During
@@ -980,10 +936,6 @@ fn cleanup_empty_artifact_dirs(local_path: &Path, start_dir: Option<&Path>) {
     }
 }
 
-// =============================================================================
-// Release Artifact Download
-// =============================================================================
-
 /// Try to download a release artifact from GitHub for the component's latest tag.
 ///
 /// Returns `Ok(Some(path))` if successful and `Ok(None)` for normal download misses
@@ -1002,7 +954,6 @@ fn try_download_release_artifact(
         return Ok(None);
     };
 
-    // Get the latest tag from the local clone (already synced by the pipeline)
     let Some(tag) = git::get_latest_tag(&component.local_path).ok().flatten() else {
         return Ok(None);
     };
@@ -1032,27 +983,20 @@ fn try_download_release_artifact(
     }
 }
 
-// =============================================================================
-// Cleanup Functions
-// =============================================================================
-
 /// Clean up build dependencies from component's local_path after successful deploy.
 /// This is a best-effort operation - failures are logged but do not fail the deploy.
 fn cleanup_build_dependencies(
     component: &Component,
     config: &DeployConfig,
 ) -> Result<Option<String>> {
-    // Skip cleanup if disabled at component level
     if !component.auto_cleanup {
         return Ok(None);
     }
 
-    // Skip cleanup if --keep-deps flag is set
     if config.keep_deps {
         return Ok(Some("skipped (--keep-deps flag)".to_string()));
     }
 
-    // Collect cleanup paths from linked extensions
     let mut cleanup_paths = Vec::new();
     if let Some(ref extensions) = component.extensions {
         for extension_id in extensions.keys() {
@@ -1134,14 +1078,11 @@ fn cleanup_build_dependencies(
 mod tests {
     use super::{
         artifact_requires_component_extract_command, cleanup_deploy_build_artifact,
-        failed_component_deploy_result, remote_version_after_deploy_effect,
-        resolve_preflight_artifact_path, should_try_download_release_artifact,
+        failed_component_deploy_result, resolve_preflight_artifact_path,
+        should_try_download_release_artifact,
     };
-    use crate::core::component::{ArtifactInput, Component, VersionTarget};
-    use crate::core::deploy::types::{DeployConfig, DeployEffect};
-    use crate::core::project::Project;
-    use crate::core::server::SshClient;
-    use std::collections::HashMap;
+    use crate::core::component::{ArtifactInput, Component};
+    use crate::core::deploy::types::DeployConfig;
     use std::process::Command;
 
     #[test]
@@ -1245,92 +1186,6 @@ mod tests {
             true,
             false,
         ));
-    }
-
-    #[test]
-    fn post_effect_version_read_uses_applied_remote_tree() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let remote_dir = temp.path().join("plugin");
-        std::fs::create_dir_all(&remote_dir).expect("remote dir");
-        std::fs::write(remote_dir.join("plugin.php"), "<?php\nVersion: 1.2.3\n")
-            .expect("remote version file");
-        let component = Component {
-            id: "plugin".to_string(),
-            remote_path: "plugin".to_string(),
-            version_targets: Some(vec![VersionTarget {
-                file: "plugin.php".to_string(),
-                pattern: Some(r"Version:\s*([0-9.]+)".to_string()),
-            }]),
-            ..Component::default()
-        };
-        let effect = DeployEffect {
-            remote_path: remote_dir.to_string_lossy().to_string(),
-            artifact_path: Some("/tmp/plugin.zip".to_string()),
-            verified: false,
-        };
-        let expected_version = "1.2.3".to_string();
-
-        let version = remote_version_after_deploy_effect(
-            &component,
-            &Project::default(),
-            temp.path().to_str().expect("base path"),
-            &local_client(),
-            Some(&effect),
-            Some(&expected_version),
-        )
-        .expect("post-effect version")
-        .expect("observed version");
-
-        assert_eq!(version, "1.2.3");
-    }
-
-    #[test]
-    fn post_effect_version_read_rejects_unobserved_configured_version_target() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let remote_dir = temp.path().join("plugin");
-        std::fs::create_dir_all(&remote_dir).expect("remote dir");
-        let component = Component {
-            id: "plugin".to_string(),
-            remote_path: "plugin".to_string(),
-            version_targets: Some(vec![VersionTarget {
-                file: "plugin.php".to_string(),
-                pattern: Some(r"Version:\s*([0-9.]+)".to_string()),
-            }]),
-            ..Component::default()
-        };
-        let effect = DeployEffect {
-            remote_path: remote_dir.to_string_lossy().to_string(),
-            artifact_path: Some("/tmp/plugin.zip".to_string()),
-            verified: false,
-        };
-        let expected_version = "1.2.3".to_string();
-
-        let error = remote_version_after_deploy_effect(
-            &component,
-            &Project::default(),
-            temp.path().to_str().expect("base path"),
-            &local_client(),
-            Some(&effect),
-            Some(&expected_version),
-        )
-        .expect_err("missing post-effect remote version should fail");
-
-        assert!(
-            error.contains("could not read remote_version from the applied tree"),
-            "unexpected error: {error}"
-        );
-    }
-
-    fn local_client() -> SshClient {
-        SshClient {
-            host: "localhost".to_string(),
-            user: "test".to_string(),
-            port: 22,
-            identity_file: None,
-            auth: None,
-            is_local: true,
-            env: HashMap::new(),
-        }
     }
 
     #[test]
