+++ src/core/deploy/mod.rs
@@ -1,3 +1,4 @@
+mod effect;
 mod execution;
 mod generated_artifacts;
 mod orchestration;
